requirements.txt
langchain
langchain-community
langchain-openai
langchain-pinecone
langchain-text-splitters

pinecone-client
elasticsearch
pypdf
python-dotenv
